requirements.txt
tensorflow==2.16.1
numpy==1.26.4
pillow==10.2.0
matplotlib==3.8.3
gunicorn==21.2.0
flask
